requirements.txt
python==3.7.7
pyresample==1.20.0
numpy==1.20.2
pyyaml==5.4.1
netcdf4==1.5.6
requests==2.25.1
xarray==0.16.2
webdavclient3==3.14.5
matplotlib==3.3.4
h5py==3.2.1
paramiko==2.7.2
scp==0.14.0
scipy==1.7.3